Fix ProgressBar redeclaration issue in GameView.swift
Resolve 'ProgressBar' redeclaration issue in `GameView.swift`.

* Add new file `ProgressBar.swift` with `ProgressBar` struct definition.
* Remove `ProgressBar` struct definition from `GameView.swift`.
* Import `ProgressBar` from `ProgressBar.swift` in `GameView.swift`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/abbayram/coffemonkey?shareId=XXXX-XXXX-XXXX-XXXX).
