+++ GameView.swift
@@ -77,28 +77,3 @@ struct GameView: View {
     }
 }
 
-struct ProgressBar: View {
-    @Binding var value: Int
-
-    var body: some View {
-        GeometryReader { geometry in
-            ZStack(alignment: .leading) {
-                Rectangle()
-                    .frame(width: geometry.size.width, height: geometry.size.height)
-                    .opacity(0.3)
-                    .foregroundColor(.gray)
-
-                Rectangle()
-                    .frame(width: min(CGFloat(self.value) / 100.0 * geometry.size.width, geometry.size.width), height: geometry.size.height)
-                    .foregroundColor(.green)
-            }
-            .cornerRadius(45.0)
-        }
-    }
-}
-
-struct GameView_Previews: PreviewProvider {
-    static var previews: some View {
-        GameView().environmentObject(GameManager(player1: Player(name: "Player 1")))
-    }
-}
